Modification stockage prix

+++ src/fr/farmeurimmo/verymc/shopgui/AmountGuiManager.java
@@ -34,8 +34,8 @@ public class AmountGuiManager implements Listener {
 						e.getInventory().getItem(22).setAmount(now);
 					}
 					ItemMeta temp = e.getInventory().getItem(22).getItemMeta();
-					Float price = BuyShopItem.pricesbuy.get(new ItemStack(Material.valueOf(e.getInventory().getItem(22).getType().toString())));
-					float totalcost = price*e.getInventory().getItem(22).getAmount();
+					Double price = BuyShopItem.pricesbuy.get(new ItemStack(Material.valueOf(e.getInventory().getItem(22).getType().toString())));
+					double totalcost = price*e.getInventory().getItem(22).getAmount();
 					temp.setLore(Arrays.asList("§6Prix d'achat: §c"+price+"$/u","§6Total: §c"+totalcost+"$"));
 					e.getInventory().getItem(22).setItemMeta(temp);
 					ItemMeta temp2 = e.getInventory().getItem(40).getItemMeta();
@@ -56,8 +56,8 @@ public class AmountGuiManager implements Listener {
 						e.getInventory().getItem(22).setAmount(now);
 					}
 					ItemMeta temp = e.getInventory().getItem(22).getItemMeta();
-					Float price = BuyShopItem.pricesbuy.get(new ItemStack(Material.valueOf(e.getInventory().getItem(22).getType().toString())));
-					float totalcost = price*e.getInventory().getItem(22).getAmount();
+					Double price = BuyShopItem.pricesbuy.get(new ItemStack(Material.valueOf(e.getInventory().getItem(22).getType().toString())));
+					double totalcost = price*e.getInventory().getItem(22).getAmount();
 					temp.setLore(Arrays.asList("§6Prix d'achat: §c"+price+"$/u","§6Total: §c"+totalcost+"$"));
 					e.getInventory().getItem(22).setItemMeta(temp);
 					ItemMeta temp2 = e.getInventory().getItem(40).getItemMeta();
@@ -66,8 +66,8 @@ public class AmountGuiManager implements Listener {
 				}
 			if(current.getType() == Material.LIME_WOOL) {
 				if(current.getAmount() == 1) {
-					Float price = BuyShopItem.pricesbuy.get(new ItemStack(Material.valueOf(e.getInventory().getItem(22).getType().toString())));
-					float amount = price;
+					Double price = BuyShopItem.pricesbuy.get(new ItemStack(Material.valueOf(e.getInventory().getItem(22).getType().toString())));
+					Double amount = price;
 					ItemStack od = e.getInventory().getItem(22);
 					int total = od.getAmount();
 					BuyShopItem.BuyOSellItemNonStack(od, (Player) e.getWhoClicked(), true, amount, total);
@@ -101,8 +101,8 @@ public class AmountGuiManager implements Listener {
 					e.getInventory().getItem(22).setAmount(now);
 				}
 				ItemMeta temp = e.getInventory().getItem(22).getItemMeta();
-				Float price = BuyShopItem.pricessell.get(new ItemStack(Material.valueOf(e.getInventory().getItem(22).getType().toString())));
-				float totalcost = price*e.getInventory().getItem(22).getAmount();
+				Double price = BuyShopItem.pricessell.get(new ItemStack(Material.valueOf(e.getInventory().getItem(22).getType().toString())));
+				double totalcost = price*e.getInventory().getItem(22).getAmount();
 				temp.setLore(Arrays.asList("§6Prix de vente: §a"+price+"$/u","§6Total: §a"+totalcost+"$"));
 				e.getInventory().getItem(22).setItemMeta(temp);
 				ItemMeta temp2 = e.getInventory().getItem(40).getItemMeta();
@@ -123,8 +123,8 @@ public class AmountGuiManager implements Listener {
 					e.getInventory().getItem(22).setAmount(now);
 				}
 				ItemMeta temp = e.getInventory().getItem(22).getItemMeta();
-				Float price = BuyShopItem.pricessell.get(new ItemStack(Material.valueOf(e.getInventory().getItem(22).getType().toString())));
-				float totalcost = price*e.getInventory().getItem(22).getAmount();
+				Double price = BuyShopItem.pricessell.get(new ItemStack(Material.valueOf(e.getInventory().getItem(22).getType().toString())));
+				double totalcost = price*e.getInventory().getItem(22).getAmount();
 				temp.setLore(Arrays.asList("§6Prix de vente: §a"+price+"$/u","§6Total: §a"+totalcost+"$"));
 				e.getInventory().getItem(22).setItemMeta(temp);
 				ItemMeta temp2 = e.getInventory().getItem(40).getItemMeta();
@@ -133,8 +133,8 @@ public class AmountGuiManager implements Listener {
 			}
 			if(current.getType() == Material.LIME_WOOL) {
 				if(current.getAmount() == 1) {
-					Float price = BuyShopItem.pricesbuy.get(new ItemStack(Material.valueOf(e.getInventory().getItem(22).getType().toString())));
-					float amount = price;
+					Double price = BuyShopItem.pricesbuy.get(new ItemStack(Material.valueOf(e.getInventory().getItem(22).getType().toString())));
+					Double amount = price;
 					ItemStack od = e.getInventory().getItem(22);
 					int total = od.getAmount();
 					BuyShopItem.BuyOSellItemNonStack(od, (Player) e.getWhoClicked(), false, amount, total);
@@ -143,7 +143,7 @@ public class AmountGuiManager implements Listener {
 			if(current.getType() == Material.GREEN_STAINED_GLASS) {
 				ItemStack od = e.getInventory().getItem(22);
 				int amountinvinv = BuyShopItem.GetAmountInInv(od, (Player) e.getWhoClicked());
-				Float price = BuyShopItem.pricesbuy.get(new ItemStack(Material.valueOf(e.getInventory().getItem(22).getType().toString())));
+				Double price = BuyShopItem.pricesbuy.get(new ItemStack(Material.valueOf(e.getInventory().getItem(22).getType().toString())));
 				if(amountinvinv >= 1) {
 					BuyShopItem.BuyOSellItemNonStack(od, (Player) e.getWhoClicked(), false, price, amountinvinv);
 				} else {
